lang: Move lang_map to lib.io for further use
